+++ tests/_utils.py
@@ -29,9 +29,7 @@ def create_test_object(
         "z": rng.uniform(0.0, 1000.0),
         "t": rng.integers(0, 1000),
         "dummy": False,
-        "states": 5,
         "label": 0,
-        "prob": 0.5,
     }
 
     obj = btrack.btypes.PyTrackObject().from_dict(data)
@@ -229,7 +227,7 @@ def full_tracker_example(
     tracker.configure(CONFIG_FILE)
     tracker.append(objects)
     tracker.volume = ((0, 1600), (0, 1200), (-1e5, 1e5))
-    tracker.track_interactive(step_size=100)
+    tracker.track(step_size=100)
     tracker.optimize()
     return tracker