+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
+*******
 Updates
-=======
+*******
+
+0.5.0
+-----
+- Factored out model specific updates from Bayesian update code
+- Added features to objects
+- Added visual update method
+
 0.4.6
 -----
 - Add deprecation warning to `track_interactive`
@@ -81,7 +89,6 @@ Updates
 - Bug fix for CSV importer
 - Added tracker.to_napari() for integration with napari track layer
 - Added change to logging verbosity when instantiating tracker
-
 
 0.3.12
 ------